pyproject.toml
[tool.poetry]
name = "contact-form"
version = "0.1.0"
description = ""
authors = ["Lucas Climenhage <Lucas.Climenhage@rwdi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-functions = "^1.17.0"
azure-communication-email = "^1.0.0"
azure-mgmt-communication = "^2.0.0"
azure-identity = "^1.15.0"
aiohttp = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
